# openapi/core-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UserController_getPublicProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/PublicProfileDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Profile công khai (tối thiểu, giữ ẩn danh) của user khác",
        "tags": ["users"]
      }
    },
...
    "/api/v1/economy/wallet": {
      "get": {
        "operationId": "EconomyController_getWallet",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/WalletDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Ví của tôi — balance + VIP (đã derive hết hạn)",
        "tags": ["economy"]
      }
    },
    "/api/v1/economy/iap/products": {
      "get": {
        "operationId": "EconomyController_listIapProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/IapProductDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Catalog gói diamond đang bán (active)",
        "tags": ["economy"]
      }
    },
    "/api/v1/economy/vip/plans": {
      "get": {
        "operationId": "EconomyController_listVipPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/VipPlanDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Catalog gói VIP đang bán (active)",
        "tags": ["economy"]
      }
    },
    "/api/v1/economy/iap/verify": {
      "post": {
        "operationId": "EconomyController_verifyIap",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyIapDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Verify receipt IAP và credit diamond — idempotent theo provider transaction id, gửi lại không credit đôi",
        "tags": ["economy"]
      }
    },
    "/api/v1/economy/vip/purchase": {
      "post": {
        "operationId": "EconomyController_purchaseVip",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PurchaseVipDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/VipPurchaseResultDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mua VIP bằng diamond — gia hạn cộng dồn nếu đang active",
        "tags": ["economy"]
      }
    },
    "/api/v1/economy/transactions": {
      "get": {
        "operationId": "EconomyController_listTransactions",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Số item tối đa mỗi trang (1-100, mặc định 20)",
            "schema": { "type": "number" }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor opaque từ `meta.nextCursor` của trang trước",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Lịch sử giao dịch — cursor pagination",
        "tags": ["economy"]
      }
    },
...
    "/api/v1/avatar/catalog": {
      "get": {
        "operationId": "AvatarController_listCatalog",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AvatarAssetDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Catalog item đang bật — giá server là nguồn sự thật",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/me/items": {
      "get": {
        "operationId": "AvatarController_listMyItems",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AvatarAssetDto" }
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Item avatar mình đang sở hữu",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/items/{assetId}/claim": {
      "post": {
        "operationId": "AvatarController_claim",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "bearer": [] }],
        "summary": "Nhận item free — idempotent",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/items/{assetId}/buy": {
      "post": {
        "operationId": "AvatarController_buy",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Bắt buộc cho mọi API có tác dụng phụ không được lặp (docs/05 § 5.4, § 5.10)",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AvatarBuyResultDto"
                    },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Mua item trả phí — trừ diamond qua Economy",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/me/equip": {
      "put": {
        "operationId": "AvatarController_equip",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EquipAvatarItemDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AvatarConfigDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Trang bị item — chỉ item đã sở hữu (docs/10 § Avatar)",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/me": {
      "get": {
        "operationId": "AvatarController_getMyAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AvatarConfigDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Avatar của chính mình — lazy-init default nếu chưa có",
        "tags": ["avatar"]
      }
    },
    "/api/v1/avatar/users/{userId}": {
      "get": {
        "operationId": "AvatarController_getAvatarOf",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data"],
                  "properties": {
                    "data": { "$ref": "#/components/schemas/AvatarConfigDto" },
                    "meta": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Xem avatar người khác — public, không cần bạn bè",
        "tags": ["avatar"]
      }
    },
...
  },
  "info": {
    "title": "Litmatch core-api",
    "description": "",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "PublicProfileDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "nickname": { "type": "string" },
          "gender": {
            "type": "string",
            "enum": ["unknown", "male", "female", "other"]
          },
          "avatarId": { "type": "string" },
          "interests": {
            "nullable": true,
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["id", "nickname", "gender", "avatarId", "interests"]
      },
...
      "WalletDto": {
        "type": "object",
        "properties": {
          "balance": {
            "type": "string",
            "example": "1200",
            "description": "Số dư diamond (chuỗi — bigint)"
          },
          "earnings": {
            "type": "string",
            "example": "0",
            "description": "Điểm quy đổi từ gift (PTS)"
          },
          "vipTier": {
            "type": "string",
            "enum": ["vip", "svip"],
            "nullable": true,
            "description": "Đã derive hết hạn — null nếu không còn active"
          },
          "vipExpiresAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": ["balance", "earnings", "vipTier", "vipExpiresAt"]
      },
      "IapProductDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "com.litmatch.diamond.100"
          },
          "provider": { "type": "string", "enum": ["apple", "google"] },
          "diamonds": {
            "type": "string",
            "example": "100",
            "description": "Số diamond nhận được (bigint dạng chuỗi)"
          }
        },
        "required": ["productId", "provider", "diamonds"]
      },
      "VipPlanDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "vip-30d" },
          "tier": { "type": "string", "enum": ["vip", "svip"] },
          "days": { "type": "number", "example": 30 },
          "priceDiamond": {
            "type": "string",
            "example": "500",
            "description": "Giá gói bằng diamond (bigint dạng chuỗi)"
          }
        },
        "required": ["id", "tier", "days", "priceDiamond"]
      },
      "VerifyIapDto": {
        "type": "object",
        "properties": {
          "provider": { "type": "string", "enum": ["apple", "google"] },
          "productId": {
            "type": "string",
            "example": "com.litmatch.diamond.100"
          },
          "payload": {
            "type": "object",
            "additionalProperties": true,
            "description": "Payload theo provider: apple {receiptData}, google {purchaseToken}; dev verifier nhận {devTransactionId}"
          }
        },
        "required": ["provider", "productId", "payload"]
      },
      "PurchaseVipDto": {
        "type": "object",
        "properties": { "planId": { "type": "string", "example": "vip-30d" } },
        "required": ["planId"]
      },
      "VipPurchaseResultDto": {
        "type": "object",
        "properties": {
          "transactionId": { "type": "string" },
          "tier": { "type": "string", "enum": ["vip", "svip"] },
          "vipExpiresAt": { "format": "date-time", "type": "string" },
          "replayed": { "type": "boolean" }
        },
        "required": ["transactionId", "tier", "vipExpiresAt", "replayed"]
      },
...
      "AvatarAssetDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slot": {
            "type": "string",
            "enum": ["base", "hair", "face", "outfit", "accessory"]
          },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "imageUrl": { "type": "string" },
          "zIndex": { "type": "number" },
          "priceDiamond": { "type": "number" }
        },
        "required": [
          "id",
          "slot",
          "code",
          "name",
          "imageUrl",
          "zIndex",
          "priceDiamond"
        ]
      },
      "AvatarBuyResultDto": {
        "type": "object",
        "properties": { "replayed": { "type": "boolean" } },
        "required": ["replayed"]
      },
      "EquipAvatarItemDto": {
        "type": "object",
        "properties": {
          "slot": {
            "type": "string",
            "enum": ["base", "hair", "face", "outfit", "accessory"]
          },
          "avatarAssetId": { "type": "string" }
        },
        "required": ["slot", "avatarAssetId"]
      },
      "AvatarConfigDto": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "layers": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AvatarAssetDto" }
          }
        },
        "required": ["userId", "layers"]
      },
...
    }
  }
}